requirements.txt
Django==2.1.7
factory_boy>=2.11.1
